small updates

+++ utils/transform.c
@@ -1604,9 +1604,9 @@ TransformSample(TRANSFORM *transform, float xv, float yv, float zv, float *px, f
       zv = gcam->mri_zind->depth-1 ;
 
     xi = nint(xv) ; yi = nint(yv) ; zi = nint(zv) ;
-    xt = (int)MRIgetVoxVal(gcam->mri_xind, xi, yi, zi, 0)*gcam->spacing ;
-    yt = (int)MRIgetVoxVal(gcam->mri_yind, xi, yi, zi, 0)*gcam->spacing ;
-    zt = (int)MRIgetVoxVal(gcam->mri_zind, xi, yi, zi, 0)*gcam->spacing ;
+    xt = nint(MRIgetVoxVal(gcam->mri_xind, xi, yi, zi, 0))*gcam->spacing ;
+    yt = nint(MRIgetVoxVal(gcam->mri_yind, xi, yi, zi, 0))*gcam->spacing ;
+    zt = nint(MRIgetVoxVal(gcam->mri_zind, xi, yi, zi, 0))*gcam->spacing ;
   }
   else
   {
